Refactoring: move aop to logging/aop
