+++ src/main/java/tech/jhipster/lite/generator/server/springboot/logging/aop/infrastructure/config/AopLoggingConfiguration.java
@@ -1,12 +1,12 @@
-package tech.jhipster.lite.generator.server.springboot.aop.logging.infrastructure.config;
+package tech.jhipster.lite.generator.server.springboot.logging.aop.infrastructure.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import tech.jhipster.lite.generator.buildtool.generic.domain.BuildToolService;
 import tech.jhipster.lite.generator.project.domain.ProjectRepository;
-import tech.jhipster.lite.generator.server.springboot.aop.logging.domain.AopLoggingDomainService;
-import tech.jhipster.lite.generator.server.springboot.aop.logging.domain.AopLoggingService;
 import tech.jhipster.lite.generator.server.springboot.common.domain.SpringBootCommonService;
+import tech.jhipster.lite.generator.server.springboot.logging.aop.domain.AopLoggingDomainService;
+import tech.jhipster.lite.generator.server.springboot.logging.aop.domain.AopLoggingService;
 
 @Configuration
 public class AopLoggingConfiguration {
