+++ src/test/java/tech/jhipster/lite/generator/server/springboot/logging/aop/infrastructure/config/AopLoggingBeanConfigurationIT.java
@@ -1,4 +1,4 @@
-package tech.jhipster.lite.generator.server.springboot.aop.logging.infrastructure.config;
+package tech.jhipster.lite.generator.server.springboot.logging.aop.infrastructure.config;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import tech.jhipster.lite.IntegrationTest;
-import tech.jhipster.lite.generator.server.springboot.aop.logging.domain.AopLoggingDomainService;
+import tech.jhipster.lite.generator.server.springboot.logging.aop.domain.AopLoggingDomainService;
 
 @IntegrationTest
 class AopLoggingBeanConfigurationIT {
