+++ tests-ci/generate.sh
@@ -34,7 +34,7 @@ if [[ $filename == 'tomcat-postgresql-ehcachejava' ]]; then
   callApi "/api/servers/spring-boot/mvc/dummy"
   callApi "/api/servers/spring-boot/databases/postgresql"
   callApi "/api/servers/spring-boot/databases/migration/liquibase"
-  callApi "/api/servers/spring-boot/aop/logging"
+  callApi "/api/servers/spring-boot/logging/aop"
   callApi "/api/servers/spring-boot/logging/logstash/tcp"
   callApi "/api/servers/spring-boot/async"
   callApi "/api/servers/spring-boot/cache/ehcache/java-configuration"
@@ -54,7 +54,7 @@ elif [[ $filename == 'tomcat-postgresql-ehcachexml' ]]; then
   callApi "/api/servers/spring-boot/mvc/security/jwt/basic-auth"
   callApi "/api/servers/spring-boot/databases/postgresql"
   callApi "/api/servers/spring-boot/databases/migration/liquibase"
-  callApi "/api/servers/spring-boot/aop/logging"
+  callApi "/api/servers/spring-boot/logging/aop"
   callApi "/api/servers/spring-boot/logging/logstash/tcp"
   callApi "/api/servers/spring-boot/async"
   callApi "/api/servers/spring-boot/cache/ehcache/xml-configuration"
@@ -74,7 +74,7 @@ elif [[ $filename == 'undertow-mysql-simplecache' ]]; then
   callApi "/api/servers/spring-boot/mvc/security/jwt/basic-auth"
   callApi "/api/servers/spring-boot/databases/mysql"
   callApi "/api/servers/spring-boot/databases/migration/liquibase"
-  callApi "/api/servers/spring-boot/aop/logging"
+  callApi "/api/servers/spring-boot/logging/aop"
   callApi "/api/servers/spring-boot/logging/logstash/tcp"
   callApi "/api/servers/spring-boot/async"
   callApi "/api/servers/spring-boot/cache/simple"
